requirements.txt
flask
werkzeug
pytesseract
opencv-python
pdf2image
reportlab
pillow
